% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedMaker.R
\name{bed12}
\alias{bed12}
\title{grl A GRangesList
bedName is bed name
uscs bed 12 format}
\usage{
bed12(grl, bedName, fixChromoNaming = FALSE)
}
\arguments{
\item{grl}{a grl}

\item{bedName}{output name}

\item{fixChromoNaming}{default FALSE}
}
\description{
grl A GRangesList
bedName is bed name
uscs bed 12 format
}
