% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{countOverlapsScore}
\alias{countOverlapsScore}
\title{Extension of countOverlaps that can use $score column}
\usage{
countOverlapsScore(query, subject, maxgap = -1L, minoverlap = 0L, type = "any")
}
\description{
Extension of countOverlaps that can use $score column
}
