% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edfClass.R
\name{create.experimenth}
\alias{create.experimenth}
\title{Create standard human experiment on server
Using hg38}
\usage{
create.experimenth(
  dir,
  exper,
  saveDir = "/export/valenfs/data/processed_data/experiment_tables_for_R/",
  types = c("bam", "bed", "wig"),
 
    txdb = "/export/valenfs/data/references/Homo_sapiens_GRCh38_110/Homo_sapiens.GRCh38.101_ensembl.gtf.db",
 
    fa = "/export/valenfs/data/references/Homo_sapiens_GRCh38_110/Homo_sapiens.GRCh38.dna.primary_assembly.fa",
  organism = "Homo sapiens",
  viewTemplate = FALSE,
  pairedEndBam = FALSE
)
}
\arguments{
\item{dir}{Which directory / directories to create experiment from}

\item{exper}{Short name of experiment, max 5 characters long}

\item{saveDir}{directory for ORFik experiments: default ("/export/valenfs/data/processed_data/experiment_tables_for_R/")}

\item{types}{Default (bam, bed, wig), which types of libraries to allow}

\item{txdb}{("/export/valenfs/data/references/Homo_sapiens_GRCh38_110/Homo_sapiens.GRCh38.101_ensembl.gtf.db")}

\item{fa}{a fasta index path ("/export/valenfs/data/references/Homo_sapiens_GRCh38_110/Homo_sapiens.GRCh38.dna.primary_assembly.fa")}

\item{organism}{"Homo sapiens"}

\item{viewTemplate}{run View() on template when finished, default (TRUE)}

\item{pairedEndBam}{logical FALSE, else TRUE, or a logical list of
TRUE/FALSE per library you see will be included (run first without and check
what order the files will come in) 1 paired end file, then two single will
be c(T, F, F)}
}
\description{
Create standard human experiment on server
Using hg38
}
