% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcp_pipeline.R
\name{difGroupAnalysis}
\alias{difGroupAnalysis}
\title{Differential group expression}
\usage{
difGroupAnalysis(cov, dir, type, grouping, test = "nonparametric")
}
\arguments{
\item{cov}{a data.table with score and feature column}

\item{type}{what is the type of score ? RNA-seq, Translational efficiency etc}

\item{grouping}{what are the category of groups ? (stop codons, start codon, motif groups etc)}

\item{test}{(defualt: nonparametric)}
}
\description{
Differential group expression
}
