% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptHelpers.R
\name{filterTranscriptsSplit}
\alias{filterTranscriptsSplit}
\title{Get filtered transcript names by groups}
\usage{
filterTranscriptsSplit(
  splitList,
  names = "txNames",
  extensions = "",
  l = 100,
  c = 100,
  t = 100,
  longestPerGene = FALSE
)
}
\arguments{
\item{splitList}{the list per element is character vector of transcripts}

\item{extension}{what is inpute name extension on parts, 1 per txdb
names = c("validMir", "validNames")
extensions = c("R", "B")}
}
\description{
Get filtered transcript names by groups
}
