% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edfLists.R
\name{getArcher16}
\alias{getArcher16}
\title{Archer 2016 Experiment info table
LIBRARY TYPES: SSU, RFP
STAGES:}
\usage{
getArcher16()
}
\description{
Archer 2016 Experiment info table
LIBRARY TYPES: SSU, RFP
STAGES:
}
