% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edfLists.R
\name{getArcher19}
\alias{getArcher19}
\title{Archer 2019 Experiment info table
LIBRARY TYPES: SSU, RNA
STAGES:}
\usage{
getArcher19()
}
\description{
Archer 2019 Experiment info table
LIBRARY TYPES: SSU, RNA
STAGES:
}
