% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edfLists.R
\name{getArcher19Bed}
\alias{getArcher19Bed}
\title{Archer 2019 Experiment info table
LIBRARY TYPES: SSU, RNA
STAGES: ?}
\usage{
getArcher19Bed()
}
\description{
Archer 2019 Experiment info table
LIBRARY TYPES: SSU, RNA
STAGES: ?
}
