% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edfLists.R
\name{getBazzini12}
\alias{getBazzini12}
\title{Bazzini 2012 Experiment info table
LIBRARY TYPES: RFP, RNA
STAGES: 64, sphere, shield (2, 4, 6)}
\usage{
getBazzini12()
}
\description{
Bazzini 2012 Experiment info table
LIBRARY TYPES: RFP, RNA
STAGES: 64, sphere, shield (2, 4, 6)
}
