% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edfLists.R
\name{getRCP5}
\alias{getRCP5}
\title{RCP-seq standard Experiment info table
LIBRARY TYPES: fractions (9-20 (-11))
shield}
\usage{
getRCP5()
}
\description{
RCP-seq standard Experiment info table
LIBRARY TYPES: fractions (9-20 (-11))
shield
}
