% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edfLists.R
\name{getRCPMZ}
\alias{getRCPMZ}
\title{RCP-seq standard Experiment info table
LIBRARY TYPES: SSU, LSU, RNA
STAGES: 64, shield}
\usage{
getRCPMZ()
}
\description{
RCP-seq standard Experiment info table
LIBRARY TYPES: SSU, LSU, RNA
STAGES: 64, shield
}
