% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edfLists.R
\name{getTCPdf}
\alias{getTCPdf}
\title{RCP-seq SSU Experiment info table
LIBRARY TYPES: SSU, RNA
STAGES: 64, sphere, shield (2, 4, 6)}
\usage{
getTCPdf(
  stage = c("64", "sphere", "shield"),
  type = c("WT", "WT", "WT"),
 
    SSU = c("/export/valenfs/projects/adam/TCP_seq/RCP_files/64cell_SSU_reps_1_2_peaks_removed_translating_filter.bam",
    "/export/valenfs/projects/adam/TCP_seq/RCP_files/sphere_SSU_reps_1_2_3_peaks_removed_translating_filter.bam",
    "/export/valenfs/projects/adam/TCP_seq/RCP_files/shield_SSU_reps_1_2_3_peaks_removed_translating_filter.bam"),
 
    RFP = c("/export/valenfs/data/processed_data/Ribo-seq/chew_2013_zebrafish/final_results/aligned_GRCz10/256Cell_trimmed.bam",
    "/export/valenfs/data/processed_data/Ribo-seq/chew_2013_zebrafish/final_results/aligned_GRCz10/Dome_trimmed.bam",
    "/export/valenfs/data/processed_data/Ribo-seq/chew_2013_zebrafish/final_results/aligned_GRCz10/Shield_trimmed.bam")
)
}
\description{
RCP-seq SSU Experiment info table
LIBRARY TYPES: SSU, RNA
STAGES: 64, sphere, shield (2, 4, 6)
}
