% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edfLists.R
\name{getTCPdfAll}
\alias{getTCPdfAll}
\title{RCP-seq SSU Experiment info table
LIBRARY TYPES: SSU, LSU, RFP, RNA
STAGES: 64, sphere, shield (2, 4, 6)}
\usage{
getTCPdfAll(
  stage = c("64", "sphere", "shield"),
  type = c("WT", "WT", "WT"),
 
    SSU = c("/export/valenfs/projects/adam/TCP_seq/RCP_files/64cell_SSU_reps_1_2_peaks_removed_translating_filter.bam",
    "/export/valenfs/projects/adam/TCP_seq/RCP_files/sphere_SSU_reps_1_2_3_peaks_removed_translating_filter.bam",
    "/export/valenfs/projects/adam/TCP_seq/RCP_files/shield_SSU_reps_1_2_3_peaks_removed_translating_filter.bam"),
 
    RFP = c("/export/valenfs/data/processed_data/Ribo-seq/chew_2013_zebrafish/final_results/aligned_GRCz10/256Cell_trimmed.bam",
    "/export/valenfs/data/processed_data/Ribo-seq/chew_2013_zebrafish/final_results/aligned_GRCz10/Dome_trimmed.bam",
    "/export/valenfs/data/processed_data/Ribo-seq/chew_2013_zebrafish/final_results/aligned_GRCz10/Shield_trimmed.bam"),
 
    RNA = c("/export/valenfs/data/processed_data/RNA-seq/lee_2013_zebrafish/total_RNA/aligned_GRCz10/WT_2hpf_Tota_mRNA.bam",
    "/export/valenfs/data/processed_data/RNA-seq/lee_2013_zebrafish/total_RNA/aligned_GRCz10/WT_4hpf_Total_mRNA.bam",
    "/export/valenfs/data/processed_data/RNA-seq/lee_2013_zebrafish/total_RNA/aligned_GRCz10/WT_6hpf_Total_mRNA_merged.bam"),
 
    LSU = c("/export/valenfs/projects/adam/TCP_seq/RCP_files/64cell_LSU_reps_1_2_peaks_removed_translating_filter.bam",
    "/export/valenfs/projects/adam/TCP_seq/RCP_files/sphere_LSU_reps_1_2_3_peaks_removed_translating_filter.bam",
    "/export/valenfs/projects/adam/TCP_seq/RCP_files/shield_LSU_reps_1_2_3_peaks_removed_translating_filter.bam")
)
}
\description{
RCP-seq SSU Experiment info table
LIBRARY TYPES: SSU, LSU, RFP, RNA
STAGES: 64, sphere, shield (2, 4, 6)
}
