% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edfLists.R
\name{getTCPNew}
\alias{getTCPNew}
\title{RCP-seq SSU 2019 mapped Experiment info table
LIBRARY TYPES: LSU
STAGES: 64, sphere, shield (2, 4, 6)}
\usage{
getTCPNew()
}
\description{
RCP-seq SSU 2019 mapped Experiment info table
LIBRARY TYPES: LSU
STAGES: 64, sphere, shield (2, 4, 6)
}
