% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slackr_support.R
\name{ggslackR}
\alias{ggslackR}
\title{Send plots to slack channel in one line}
\usage{
ggslackR(
  plot = last_plot(),
  width = 300,
  height = 150,
  config_file = "/export/valenfs/projects/Hakon/slackr.config"
)
}
\arguments{
\item{plot}{last_plot(), a ggplot sent to Rstudio frame viewer}

\item{width}{in mm, default (300)}

\item{config_file}{path to you config file}

\item{heigth}{in mm, default (150)}
}
\description{
You need a config file:
To make it, follow:
https://github.com/hrbrmstr/slackr
}
