% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{GRToWig}
\alias{GRToWig}
\title{Export GRanges object as 2 wig files}
\usage{
GRToWig(gr, outputPrefix = p(mainFolder, "/test_"))
}
\arguments{
\item{gr}{GRanges object}

\item{outputPrefix}{prefix and folder for output,
will add forward and reverse}
}
\description{
Split into +/- strand
}
