% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcp_pipeline.R
\name{heatMapL}
\alias{heatMapL}
\title{heatmap
Old default path:  = p(mainFolder, "/tcp_plots/mir430/hm_")}
\usage{
heatMapL(
  region,
  tx,
  df,
  outdir,
  scores = "sum",
  upstream,
  downstream,
  zeroPosition = upstream,
  acLen = NULL,
  legendPos = "right",
  colors = NULL,
  addFracPlot = TRUE,
  location = "TIS",
  shifting = NULL,
  skip.last = FALSE,
  format = ".png"
)
}
\arguments{
\item{outdir}{directory to save to (default: NULL, no saving)}

\item{scores}{scoring function (default: c("sum", "zscore")),
see ?coverageScorings for possible scores.}

\item{colors}{Which colors to use, default (skyblue4)}
}
\description{
heatmap
Old default path:  = p(mainFolder, "/tcp_plots/mir430/hm_")
}
