% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcp_pipeline.R
\name{heatMapRegion}
\alias{heatMapRegion}
\title{Create heatmap of specified region}
\usage{
heatMapRegion(
  df,
  region = "TIS",
  outdir = paste0(dirname(df$filepath[1]), "/QC_STATS/heatmaps/"),
  scores = c("transcriptNormalized", "sum"),
  cage = NULL,
  format = ".png"
)
}
\arguments{
\item{outdir}{directory to save to (default: NULL, no saving)}

\item{scores}{scoring function (default: c("sum", "zscore")),
see ?coverageScorings for possible scores.}
}
\description{
Simplified input space for easier abstraction
}
