% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edfClass.R
\name{list.experiments}
\alias{list.experiments}
\title{List current experiment available}
\usage{
list.experiments(
  dir = "/export/valenfs/data/processed_data/experiment_tables_for_R/",
  pattern = "*",
  libtypeExclusive = NULL
)
}
\arguments{
\item{dir}{directory for ORFik experiments: default ("/export/valenfs/data/processed_data/experiment_tables_for_R/")}

\item{pattern}{allowed patterns in experiment file name: default ("*", all)}

\item{libtypeExclusive}{search for experiments with exclusivly this libtype, default (NULL, all)}
}
\description{
Will only search .csv extension, also exclude any experiment with the word template.
}
