% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{pipelineCluster}
\alias{pipelineCluster}
\title{set up cluster for pipeline}
\usage{
pipelineCluster(maxCores = NULL, reset = FALSE, outfile = NULL)
}
\arguments{
\item{maxCores}{what is max cores, if Null set to half of available cores}

\item{reset}{reset cluster, use only if already existing}

\item{outfile}{log outputs}
}
\description{
Cluster object saved as cl
}
