% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcp_pipeline.R
\name{plotHelper}
\alias{plotHelper}
\title{Helper function for coverage plots}
\usage{
plotHelper(
  coverage,
  df,
  outdir,
  scores,
  returnCoverage = FALSE,
  title = "coverage metaplot",
  colors = c("skyblue4", "orange"),
  plotFunction = "windowCoveragePlot"
)
}
\arguments{
\item{outdir}{directory to save to (default: NULL, no saving)}

\item{scores}{scoring function (default: c("sum", "zscore")),
see ?coverageScorings for possible scores.}

\item{returnCoverage}{(defualt: FALSE), return the ggplot object (TRUE)
or NULL (FALSE).}

\item{title}{Title to give plot}

\item{colors}{Which colors to use, default (skyblue4)}

\item{plotFunction}{Which plot function, default: windowCoveragePlot}
}
\value{
NULL (or ggplot object if returnCoverage is TRUE)
}
\description{
Helper function for coverage plots
}
