% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edfClass.R
\name{read.experimentl}
\alias{read.experimentl}
\title{Read standard experiment on server}
\usage{
read.experimentl(
  relPath,
  dir = "/export/valenfs/data/processed_data/experiment_tables_for_R/",
  expInVarName = FALSE
)
}
\arguments{
\item{relPath}{name of experiment, run list.experiments() for candidates}

\item{dir}{directory for ORFik experiments: default ("/export/valenfs/data/processed_data/experiment_tables_for_R/")}

\item{expInVarName}{keep experiment name in variable names? defualt (FALSE)}
}
\description{
Read standard experiment on server
}
