% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edfClass.R
\name{readBamNew}
\alias{readBamNew}
\title{Try to optimze bam reader}
\usage{
readBamNew(path, samtoolsPath = "/Home/ii/hakontj/bin/samtools-1.9/samtools")
}
\arguments{
\item{path}{a path to bam file}

\item{samtoolsPath}{If not samtools in global scope, give full path here to samtools binary}
}
\value{
a GAlignment object of bamfile
}
\description{
Extract fields 3,4,6 and 7
3: chromosome
4: left most position,
}
