% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcp_pipeline.R
\name{regionWindow}
\alias{regionWindow}
\title{Meta coverage of region}
\usage{
regionWindow(
  region,
  tx,
  df = getTCPdf(),
  outdir = p(mainFolder, "/tcp_plots/TIS_region/target_"),
  upstream = 75,
  downstream = 74,
  scores = c("sum", "zscore"),
  colors = c("skyblue4", "orange"),
  allTogether = TRUE,
  title = "cds metacoverage",
  dfr = NULL
)
}
\arguments{
\item{dfr}{RNA seq ORFik experiment to normalize against}
}
\description{
Meta coverage of region
}
