% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcp_pipeline.R
\name{regionWindowAll}
\alias{regionWindowAll}
\title{Meta coverage of region using two groups}
\usage{
regionWindowAll(
  one,
  two,
  tx,
  df = getTCPdf(),
  outdir,
  upstream = 75,
  downstream = 74,
  scores = c("sum", "zscore"),
  seperations = c("lowly", "highly"),
  title = "cds metacoverage",
  colors = c("skyblue4", "orange"),
  allTogether = TRUE,
  dfr = NULL
)
}
\arguments{
\item{dfr}{RNA seq ORFik experiment to normalize positions against}
}
\description{
Meta coverage of region using two groups
}
