% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcp_pipeline.R
\name{removeBadTxByRegion}
\alias{removeBadTxByRegion}
\title{Filter out transcript by a median filter}
\usage{
removeBadTxByRegion(
  tx,
  reads,
  upstream = NULL,
  downstream = NULL,
  median_multiplier = "0.99",
  extension = max(0, upstream + 1),
  min_cutoff = "0.999"
)
}
\arguments{
\item{min_cutoff}{either a quantile if input is string[0-1], like "0.99", or numeric value if input is numeric.}
}
\description{
Filter out transcript by a median filter
}
