% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{removeNonFinite}
\alias{removeNonFinite}
\title{remove non finite in data.table}
\usage{
removeNonFinite(DT, replacement = 0)
}
\arguments{
\item{DT}{a data.table}

\item{replacement}{(0) what should the non finite values be changed to ?}
}
\description{
Removes inf, NA, and NaN values
}
