% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edfClass.R
\name{save.experimentl}
\alias{save.experimentl}
\title{Save template experiment to disc}
\usage{
save.experimentl(
  temp,
  saveDir = "/export/valenfs/data/processed_data/experiment_tables_for_R/"
)
}
\arguments{
\item{temp}{a temperary ORFik experiment template}

\item{saveDir}{directory for ORFik experiments: default ("/export/valenfs/data/processed_data/experiment_tables_for_R/")}
}
\description{
Save template experiment to disc
}
