% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedExperimentHelpers.R
\name{SEdif}
\alias{SEdif}
\title{Get difference between wild type and target variant
Per library type like RNA-seq or Ribo-seq.}
\usage{
SEdif(dt, df)
}
\arguments{
\item{dt}{a data.table of counts or fpkm etc.}
}
\value{
a data.table with differences
}
\description{
Get difference between wild type and target variant
Per library type like RNA-seq or Ribo-seq.
}
