% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcp_pipeline.R
\name{splitMetacoverage}
\alias{splitMetacoverage}
\title{Split metacoverage by a splitList}
\usage{
splitMetacoverage(
  region,
  splitList,
  df,
  outdir,
  scores = ifelse(!is.null(dfr), c("sum", "zscore"), "sum"),
  colors = c("skyblue4", "orange"),
  allTogether = TRUE,
  title = "cds metacoverage",
  dfr = NULL
)
}
\arguments{
\item{splitList}{character vector where each unique group will be an output}
}
\description{
Split metacoverage by a splitList
}
