% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptHelpers.R
\name{splitRegions}
\alias{splitRegions}
\title{Split regions into groups
Assigns to .GlobalEnv all variables needed}
\usage{
splitRegions(
  parts = c("mrna", "leaders", "cds", "trailers"),
  splitList,
  extensions = "",
  splitExt = ""
)
}
\arguments{
\item{splitList}{the list per element is character vector of transcripts to keep}

\item{splitExt}{what is output name extension on parts}

\item{extension}{what is inpute name extension on parts}
}
\description{
Split regions into groups
Assigns to .GlobalEnv all variables needed
}
