% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAR.R
\name{STAR.folder}
\alias{STAR.folder}
\title{Align data with STAR}
\usage{
STAR.folder(
  input.dir,
  output.dir,
  index.dir,
  paired.end = "no",
  steps = "tr-ge",
  adapter.sequence = "auto",
  min.length = 15,
  trim.front = 3,
  alignment.type = "Local",
  max.cpus = 90,
  include.subfolders = "n",
 
    script = "/export/valenfs/projects/Pipelines/STAR_Aligner/RNA_Align_pipeline_folder.sh -h"
)
}
\description{
Align data with STAR
}
