% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAR.R
\name{STAR.index}
\alias{STAR.index}
\title{Create STAR genome index}
\usage{
STAR.index(
  GTF,
  fasta,
  output.dir,
  rrna = NULL,
  trna = NULL,
  ncrna = NULL,
  phix = NULL,
  script = "/export/valenfs/projects/Pipelines/STAR_Aligner/STAR_MAKE_INDEX.sh"
)
}
\description{
Used as reference when aligning data
}
