% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcp_pipeline.R
\name{tcpHeatMap_single}
\alias{tcpHeatMap_single}
\title{Heatmap of single library}
\usage{
tcpHeatMap_single(
  region,
  tx,
  reads,
  outdir,
  scores = "sum",
  upstream,
  downstream,
  zeroPosition = upstream,
  returnCoverage = FALSE,
  acLen = NULL,
  legendPos = "right",
  colors = NULL,
  addFracPlot = TRUE,
  location = "start site",
  shifting = NULL,
  skip.last = FALSE
)
}
\arguments{
\item{outdir}{directory to save to (default: NULL, no saving)}

\item{scores}{scoring function (default: c("sum", "zscore")),
see ?coverageScorings for possible scores.}

\item{colors}{Which colors to use, default (skyblue4)}
}
\description{
Heatmap of single library
}
