% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcp_pipeline.R
\name{tRNA_stats}
\alias{tRNA_stats}
\title{Find ratio of iMet to met amino acids
Also get basic tRNA statistics}
\usage{
tRNA_stats(
  fa,
  bamFiles,
 
    trna = "/export/valenfs/data/references/Zv10_zebrafish/tRNAscan-SE-2.0/GRcZ10_tRNA_scan_output.fa"
)
}
\arguments{
\item{fa}{location of fasta genome file which tRNA aligned to, or a biostring object of preloaded}

\item{bamFiles}{paths to bamFiles matched to fa. A character vector of path to bam files aligned to fa.}

\item{trna}{the fasta tRNA from trnaScan ?}
}
\value{
a data.frame of imet and met content
}
\description{
Find ratio of iMet to met amino acids
Also get basic tRNA statistics
}
