% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{updateORFik}
\alias{updateORFik}
\title{Update ORFik version}
\usage{
updateORFik(branch = "master", user = "Roleren")
}
\arguments{
\item{branch}{default: "master"}

\item{user}{default: Roleren}
}
\description{
Update ORFik version
}
