% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edfClass.R
\name{create.experimentl}
\alias{create.experimentl}
\title{Create standard zebrafish experiment on server
Using Z10}
\usage{
create.experimentl(
  dir,
  exper,
  saveDir = "/export/valenfs/data/processed_data/experiment_tables_for_R/",
  types = c("bam", "bed", "wig"),
 
    txdb = "/export/valenfs/data/references/Zv10_zebrafish/Danio_rerio.GRCz10.81_chr.gtf.db",
  fa = "/export/valenfs/data/references/Zv10_zebrafish/Danio_rerio.GRCz10.fa",
  organism = "Danio rerio",
  viewTemplate = FALSE,
  pairedEndBam = FALSE
)
}
\arguments{
\item{dir}{Which directory / directories to create experiment from}

\item{exper}{Short name of experiment, max 5 characters long}

\item{saveDir}{directory for ORFik experiments: default ("/export/valenfs/data/processed_data/experiment_tables_for_R/")}

\item{types}{Default (bam, bed, wig), which types of libraries to allow}

\item{txdb}{("/export/valenfs/projects/uORFome/Annotations/Zebrafish/zebrafish_GRCh10_81.gtf.db")}

\item{fa}{a fasta index path ("/export/valenfs/projects/uORFome/Annotations/Zebrafish/zebrafish_GRCh10_81.gtf.db")}

\item{organism}{Danio rerio}

\item{viewTemplate}{run View() on template when finished, default (TRUE)}

\item{pairedEndBam}{logical FALSE, else TRUE, or a logical list of
TRUE/FALSE per library you see will be included (run first without and check
what order the files will come in) 1 paired end file, then two single will
be c(T, F, F)}
}
\description{
Create standard zebrafish experiment on server
Using Z10
}
