% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcp_pipeline.R
\name{plotHelper}
\alias{plotHelper}
\title{Helper function for coverage plots}
\usage{
plotHelper(
  coverage,
  df,
  outdir,
  scores,
  returnCoverage = FALSE,
  title = "coverage metaplot",
  colors = c("skyblue4", "orange"),
  plotFunction = "windowCoveragePlot"
)
}
\arguments{
\item{outdir}{a character path to save file as: not just directory,
but full name.}

\item{scores}{character vector, default "sum",
either of zscore, transcriptNormalized, sum, mean, median, ..
see ?coverageScorings for info and more alternatives.}

\item{returnCoverage}{(defualt: FALSE), return the ggplot object (TRUE)
or NULL (FALSE).}

\item{title}{Title to give plot}

\item{colors}{character vector, default: "default", this gives you:
c("white", "yellow2", "yellow3", "lightblue", "blue", "navy"),
do "high" for more high contrasts, or specify your own colors.}

\item{plotFunction}{Which plot function, default: windowCoveragePlot}
}
\value{
NULL (or ggplot object if returnCoverage is TRUE)
}
\description{
Helper function for coverage plots
}
